+++ actions/log_action/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0
+
+- chore: bump dart version to 3.0.0 ([#250](https://github.com/wolfenrain/fluttium/issues/250))
+- chore: bump fluttium version ([#259](https://github.com/wolfenrain/fluttium/issues/259))
+
 # 0.1.0+1
 
 - feat: add simple log action ([#146](https://github.com/wolfenrain/fluttium/issues/146))