+++ packages/fluttium/CHANGELOG.md
@@ -1,3 +1,16 @@
+# 0.1.0
+
+- feat!: migrate to Flutter v3.10.0 ([#244](https://github.com/wolfenrain/fluttium/issues/244))
+- chore: bump dart version to 3.0.0 ([#250](https://github.com/wolfenrain/fluttium/issues/250))
+
+# 0.1.0-dev.4
+
+- feat: support pixel ratio for the screenshot action ([#229](https://github.com/wolfenrain/fluttium/issues/229))
+
+# 0.1.0-dev.3
+
+- feat: implement scroll and swipe actions ([#219](https://github.com/wolfenrain/fluttium/issues/219))
+
 # 0.1.0-dev.2
 
 - feat: implement a `Wait` action ([#186](https://github.com/wolfenrain/fluttium/issues/186))