+++ packages/fluttium_driver/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.1.1
+
+- fix: use correct fluttium constraints ([#261](https://github.com/wolfenrain/fluttium/issues/261))
+
+# 0.1.0
+
+- feat!: migrate to Flutter v3.10.0 ([#246](https://github.com/wolfenrain/fluttium/issues/246))
+- chore: bump dart version to 3.0.0 ([#250](https://github.com/wolfenrain/fluttium/issues/250))
+
 # 0.1.0-dev.3
 
 - fix: support path for actions on windows ([#193](https://github.com/wolfenrain/fluttium/issues/193))