+++ packages/fluttium_protocol/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.0
+
+- chore: bump dart version to 3.0.0 ([#250](https://github.com/wolfenrain/fluttium/issues/250))
+
 # 0.1.0-dev.1
 
 - feat: add support for custom actions ([#112](https://github.com/wolfenrain/fluttium/issues/112))